backend/requirements.txt
# Google ADK and Core
google-adk>=1.11.0
google-api-python-client>=2.100.0
google-search-results>=2.4.2

# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Data Processing (Optimized versions)
pandas>=2.1.0
numpy>=1.24.0
polars>=0.20.0  # High-performance alternative to pandas

# Machine Learning (Latest optimized versions)
scikit-learn>=1.4.0
xgboost>=2.0.3
lightgbm>=4.2.0
catboost>=1.2.2
optuna>=3.5.0

# Utilities
joblib>=1.3.2
psutil>=5.9.0  # System performance monitoring
cachetools>=5.3.0  # Caching for performance
asyncio-pool>=0.6.0  # Async task pooling

# Development and Monitoring
rich>=13.7.0  # Beautiful terminal output
loguru>=0.7.2  # Advanced logging
